Display pictures from all users. Arrange user pictures better.
The user pictures are displayed in a column of its own on the right
part of the screen, in a carousel to make them easy to go through
without taking up much space, since the user is more likely to know his
pictures already.

Pictures from all users are displayed in the middle of the screen in a
row of thumbnails.